Bugfix mkdir_p: Don't call os.makedirs if path is empty
In line 1810, `mkdir_p` is called with the dirname of the outfile. If the outfile is to be created in the current working directory, `os.path.dirname()` will return the empty string. Then, `os.makedirs()` in line 1159 will raise an error: `OSError: [Errno 2] No such file or directory: ''`.

The proposed change checks for the empty string and returns without trying to create the directory.

To reproduce the bug please call js-beautify with a file in the current working directory, e.g. `js-beautify -r some-file.js` and ensure that js-beautify has indeed something to beautify, otherwise the file won't be touched.